Add server-Makefile to repo
The server-Makefile is a template for files that overwrite student
Makefile during a submission. It only needs the following targets:
  release
  debug
  all (builds release and debug)
  identifier
  clean

To make this useful, Makefile was edited and reorganized to minimize the
diff report between the two files. This will make it less painful when
changes are needed, to make sure they are done in the proper file or
files.