Reorder student Makefile targets so debug is first
The student Makefile is for development and the more common usage is to
build and debug, so the debug target has been moved to the top allowing
for both `make` and `make debug` to do the "most useful thing".

The documentation has been updated to reflect the changes made to the
Makefile.